pyproject.toml
[project]
name = "llm-beginner-guide"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aider-chat>=0.78.0",
    "openai>=1.66.3",
    "rich>=13.9.4",
]
